smooth moving character

+++ Assets/Skript/Skin.cs
@@ -4,12 +4,12 @@
 public class Skin : ScriptableObject
 {
     [SerializeField] private Mesh _mesh;
-    [SerializeField] private float _cost;
+    [SerializeField] private int _cost;
     
     [SerializeField]private bool _isSold = false;
     
     public Mesh SkinMesh => _mesh;
-    public float Cost => _cost;
+    public int Cost => _cost;
     public bool IsSold => _isSold;
     
     public void Buy()
